requirements.txt
streamlit==1.33.0
scikit-learn==1.7.2
pandas==2.2.1
numpy==1.26.4
lime==0.2.0.1
joblib==1.4.2
matplotlib==3.8.3
requests
beautifulsoup4
lxml
sentence-transformers==2.7.0
torch
fpdf2==2.7.9
Pillow
pyzbar
